cardrecognition/requirements.txt
opencv-python
numpy
python-dotenv
openai